+++ src/main/java/com/mumulcom/mumulcom/src/reply/service/ReplyService.java
@@ -95,7 +95,7 @@ public class ReplyService {
      * yeji
      * 전체 답변 조회 API
      */
-    public List<GetReplyRes> getReplyList(int questionIdx, int userIdx) throws BaseException {
+    public List<GetReplyRes> getReplyList(Long questionIdx, Long userIdx) throws BaseException {
         try {
             List<GetReplyRes> getReplyRes = replyDao.getReplyList(questionIdx, userIdx);
             return getReplyRes;
